feat: implement `init` command (#5)
This PR implements the `init` command in a procedural style.

`init` command, if `project_name` provided, creates a new project
directory with it's corresponding `src/` and `include/` directories, as
well as `nimbus.toml` configuration file. In case the `project_name` is
not provided, it uses the current directory as the project directory.

Closes #3